+++ src/Memory/RoomStats.js
@@ -4,80 +4,43 @@
 * @param {Object} room
 * @param {string} target
 */
-function CheckController(room, target) {
-    if(target == "State") {
-        if (room.controller) return true;
-        else return false;
-    } else if(target == "IsMy") {
-        if (room.controller && room.controller.my) return true;
-        else if (room.controller && !room.controller.my) return false;
-        else return null;
-    } else if(target == "Reserved") {
-        if (room.controller && room.controller.reservation) return room.controller.reservation;
-        else return false;
-    } else if(target == "Level") {
-        if (room.controller && room.controller.my && room.controller.level) return room.controller.level;
-        else return null;
-    } else if(target == "Progress") {
-        if (room.controller && room.controller.my) return room.controller.progress / room.controller.progressTotal * 100;
-        else return null;
-    } else if(target == "Downgrade") {
-        if (room.controller && room.controller.my) return room.controller.ticksToDowngrade;
-        else return null;
-    } else if(target == "Sign") {
-        if (room.controller && room.controller.my && room.controller.sign) return room.controller.sign;
-        else return null;
-    } else if(target == "SafeMode") {
-        if (room.controller && room.controller.my && room.controller.safeMode) return room.controller.safeMode;
-        else return null;
-    } else if(target == "SafeModeAvailable") {
-        if (room.controller && room.controller.my && room.controller.safeModeAvailable) return room.controller.safeModeAvailable;
-        else return null;
-    } else if(target == "SafeModeCooldown") {
-        if (room.controller && room.controller.my && room.controller.safeModeCooldown) return room.controller.safeModeCooldown;
-        else return null;
-    }
-}
+const CheckRoom = {
+    Storage(room) {
+        let owners = [];
+        let info = {};
 
-/**
-* @param {Object} room
-* @param {string} target
-*/
-function checkStorage(room, target) {
-    if(target == "State") {
-        if (room.storage) return true;
-        else return false;
-    } else if(target == "CapacityUsed") {
-        if (room.storage) return room.storage.store.getUsedCapacity();
-        else return null;
-    } else if(target == "CapacityFree") {
-        if (room.storage) return room.storage.store.getFreeCapacity();
-        else return null;
-    } else if(target == "CapacityEnergy") {
-        if (room.storage) return room.storage.store[RESOURCE_ENERGY];
-        else return null;
-    }
-}
+        let storage = false;
+        let energyBalance = null;
 
-/**
-* @param {Object} room
-* @param {string} target
-*/
-function checkRoom(room, target) {
-    let info = null;
-    let owners = [];
+        let energyCapacity = null;
+        let usedCapacity = null;
+        let freeCapacity = null;
 
-    if(target == "EnergyBalance") {
         if (room.storage) {
-            if (room.terminal) {
-                if (room.storage.store[RESOURCE_ENERGY] + room.terminal[RESOURCE_ENERGY] > 200000) return true;
-                else return false
-            } else {
-                if (room.storage.store[RESOURCE_ENERGY] > 100000) return true;
-                else return false
-            }
-        } else return null;
-    } else if(target == "HostileCreeps") {
+            storage = true;
+            if (room.storage.store[RESOURCE_ENERGY] + room.terminal[RESOURCE_ENERGY] > 200000) energyBalance = true;
+            else energyBalance = false;
+
+            energyCapacity = room.storage.store[RESOURCE_ENERGY];
+            usedCapacity = room.storage.store.getUsedCapacity();
+            freeCapacity = room.storage.store.getFreeCapacity();
+        }
+
+        info = {
+            State: storage,
+            EnergyBalance: energyBalance,
+            Energy: energyCapacity,
+            UsedCapacity: usedCapacity,
+            FreeCapacity: freeCapacity
+        }
+
+        return info;
+    },
+
+    HostileCreeps(room) {
+        let owners = [];
+        let info = {};
+
         const HostileCreeps = room.find(FIND_HOSTILE_CREEPS, {
             filter: (creep) => {
                 return (!Memory.Friends.includes(creep.owner.username));
@@ -89,88 +52,195 @@ function checkRoom(room, target) {
                 if (!owners.includes(HostileCreeps[i].owner.username)) owners.push(HostileCreeps[i].owner.username);
             }
             info = {
-                HostileState:true,
+                State:true,
                 Amount:HostileCreeps.length,
                 Owners:owners,
             }
         } else {
             info = {
-                HostileState:false,
+                State:false,
                 Amount:null,
                 Owners:null,
             }
         }
 
         return info;
-    } else if(target == "EnergySources") {
+    },
+
+    Sources(room) {
+        let info = {};
+
+        let sourcesAmount = 0;
+        let firstSource = null;
+        let secondSource = null;
+
         const sources = room.find(FIND_SOURCES);
-        return sources.length;
-    } else if(target == "Source1") {
-        const sources = room.find(FIND_SOURCES);
-        return sources[0].id;
-    } else if(target == "Source2") {
-        const sources = room.find(FIND_SOURCES);
-        if (sources[1]) return sources[1].id;
-        else return false;
-    } else if(target == "RoomMineral") {
+
+        if (sources[0]) firstSource = sources[0].id;
+        if (sources[1]) secondSource = sources[1].id;
+
+        if (sources.length > 0) sourcesAmount = sources.length
+
+        info = {
+            Amount: sourcesAmount,
+            FirstSource: firstSource,
+            SecondSource: secondSource
+        }
+
+        return info;
+    },
+
+    Spawns(room) {
+        let info = {};
+
+        let firstSpawn = null;
+        let secondSpawn = null;
+        let thirdSpawn = null;
+        let amountSpawns = 0;
+
+        const spawns = room.find(FIND_MY_SPAWNS);
+
+        if (spawns.length > 0) {
+            amountSpawns = spawns.length
+
+            if (spawns[0]) firstSpawn = spawns[0].id;
+            if (spawns[1]) secondSpawn = spawns[1].id;
+            if (spawns[2]) thirdSpawn = spawns[2].id;
+        }
+
+        info = {
+            Amount: amountSpawns,
+            FirstSpawn: firstSpawn,
+            SecondSpawn: secondSpawn,
+            ThirdSpawn: thirdSpawn
+        }
+
+        return info;
+    },
+
+    Minerals(room) {
+        let info = {};
+
+        let sourceInRoom = null;
+        let extractorInRoom = false;
+        let mineralRegeneration = null;
+
         const source = room.find(FIND_MINERALS);
-        return source[0].id;
-    } else if(target == "MineralRegeneration") {
-        const source = room.find(FIND_MINERALS);
-        return source[0].ticksToRegeneration;
-    } else if(target == "Nuke") {
-        return false;
-    } else if (target == "Spawn1") {
-        const spawns = room.find(FIND_MY_SPAWNS);
-        if (spawns[0]) return spawns[0].id;
-        else return null;
-    } else if (target == "Spawn2") {
-        const spawns = room.find(FIND_MY_SPAWNS);
-        if (spawns[1]) return spawns[1].id;
-        else return null;
-    } else if (target == "Spawn3") {
-        const spawns = room.find(FIND_MY_SPAWNS);
-        if (spawns[2]) return spawns[2].id;
-        else return null;
-    } else if (target == 'Extractor') {
-        const extractor = room.find(FIND_STRUCTURES, {
-            filter: (strc) => {
-                return strc.type == 'extractor';
+
+        if (source[0]) {
+            sourceInRoom = source[0].id;
+            mineralRegeneration = source[0].ticksToRegeneration;
+
+            const extractor = room.find(FIND_STRUCTURES, {
+                filter: (strc) => {
+                    return strc.type == 'extractor';
+                }
+            });
+
+            if (extractor.length > 0) extractorInRoom = extractor[0].id;
+        }
+
+        info = {
+            Source: sourceInRoom,
+            MineralRegeneration: mineralRegeneration,
+            Extractor: extractorInRoom
+        }
+
+        return info;
+    },
+
+    Terminal(room) {
+        let info = {};
+
+        let terminalInRoom = false;
+        let energyCapacity = null;
+        let usedCapacity = null;
+        let freeCapacity = null;
+
+        if (room.terminal) {
+            terminalInRoom = true;
+            energyCapacity = room.terminal.store[RESOURCE_ENERGY];
+            usedCapacity = room.terminal.store.getUsedCapacity();
+            freeCapacity = room.terminal.store.getFreeCapacity();
+        }
+
+        info = {
+            State: terminalInRoom,
+            Energy: energyCapacity,
+            UsedCapacity: usedCapacity,
+            FreeCapacity: freeCapacity
+        }
+
+        return info;
+    },
+
+    Controller(room) {
+        let info = {};
+
+        let state = false;
+        let isMy = null;
+        let reserved = null;
+        let level = null;
+        let progress = null;
+        let downgrade = null;
+        let sign = null;
+        let safeMode = null;
+        let safeModeAvailable = null;
+        let safeModeCooldown = null;
+
+        if (room.controller) {
+            state = true;
+            if (room.controller.my) {
+                isMy = true;
+                level = room.controller.level;
+                progress = room.controller.progress / room.controller.progressTotal * 100;
+                downgrade = room.controller.ticksToDowngrade;
+                if (room.controller.sign) sign = room.controller.sign;
+                if (room.controller.safeMode) safeMode = room.controller.safeMode;
+                if (room.controller.safeModeAvailable) safeModeAvailable = room.controller.safeModeAvailable;
+                if (room.controller.safeModeCooldown) safeModeCooldown = room.controller.safeModeCooldown;
             }
-        });
-        if (extractor.length > 0) return true;
-        else return false;
-    }
-}
+            else if (room.controller.reservation) reserved = room.controller.reservation;
+        }
 
-/**
-* @param {Object} room
-* @param {string} target
-*/
-function checkTerminal(room, target) {
-    if(target == "State") {
-        if (room.terminal) return true;
-        else return false;
-    } else if(target == "CapacityUsed") {
-        if (room.terminal) return room.terminal.getUsedCapacity();
-        else return null;
-    } else if(target == "CapacityFree") {
-        if (room.terminal) return room.terminal.getFreeCapacity();
-        else return null;
-    } else if(target == "CapacityEnergy") {
-        if (room.terminal) return room.terminal.store[RESOURCE_ENERGY];
-        else return null;
-    }
-}
+        info = {
+            State: state,
+            IsMy: isMy,
+            Reserved: reserved,
+            Level: level,
+            Progress: progress,
+            Downgrade: downgrade,
+            Sigh: sign,
+            SafeMode: safeMode,
+            SafeModeAvailable: safeModeAvailable,
+            SafeModeCooldown: safeModeCooldown,
+        }
 
-/**
-* @param {string} target
-*/
-function checkPlayer(target) {
-    if(target == "Credits") {
-        return null;
-    } else if(target == "GCL") {
-        return null;
+        return info;
+    },
+
+    OtherInfo(room) {
+        let info = {};
+
+        let constructionSites = [];
+        let amountConstructionSites = 0;
+
+        const cs = room.find(FIND_CONSTRUCTION_SITES);
+
+        if (cs.length > 0) {
+            constructionSites = cs;
+            amountConstructionSites = cs.length;
+        }
+
+        info = {
+            ConstructionSites: {
+                Amount: amountConstructionSites,
+                List: constructionSites,
+            },
+            Nuke: false,
+        }
+
+        return info;
     }
 }
 
@@ -183,53 +253,14 @@ function RoomStats() {
         Info = {
             RoomName:room.name,
             Room:{
-                EnergyBalance:checkRoom(room, "EnergyBalance"),
-                HostileCreeps:checkRoom(room, "HostileCreeps"), // bool
-                Sources:{
-                    Amount:checkRoom(room, "EnergySources"),
-                    FirstSource:checkRoom(room, "Source1"),
-                    SecondSource:checkRoom(room, "Source2"),
-                },
-                Spawns:{
-                    Amount:checkRoom(room, "Spawns:Amount"),
-                    Spawn1:checkRoom(room, "Spawn1"),
-                    Spawn2:checkRoom(room, "Spawn2"),
-                    Spawn3:checkRoom(room, "Spawn3"),
-                },
-                Mineral:{
-                    RoomMineral:checkRoom(room, "RoomMineral"),
-                    Extractor:checkRoom(room, "Extractor"),
-                    MineralRegeneration:checkRoom(room, "MineralRegeneration"),
-                },
-                ConstructionSites:{
-                    Amount:checkRoom(room, "ConstructionSites:Amount"),
-                    List:checkRoom(room, "ConstructionSites:List"),
-                },
-                Nukes:checkRoom(room, "Nuke"),
-                Controller:{
-                    ControllerState:CheckController(room, "State"),
-                    ControllerIsMy:CheckController(room, "IsMy"),
-                    ControllerReserved:CheckController(room, "Reserved"),
-                    ControllerLevel:CheckController(room, "Level"),
-                    ControllerProgress:CheckController(room, "Progress"),
-                    ControllerDowngrade:CheckController(room, "Downgrade"),
-                    ControllerSign:CheckController(room, "Sign"),
-                    ControllerSafeMode:CheckController(room, "SafeMode"),
-                    ControllerSafeModeAvailable:CheckController(room, "SafeModeAvailable"),
-                    ControllerSafeModeCooldown:CheckController(room, "SafeModeCooldown"),
-                },
-                Terminal:{
-                    TerminalState:checkTerminal(room, "State"),
-                    TerminalCapacityUsed:checkTerminal(room, "CapacityUsed"),
-                    TerminalCapacityFree:checkTerminal(room, "CapacityFree"),
-                    TerminalCapacityEnergy:checkTerminal(room, "CapacityEnergy"),
-                },
-                Storage:{
-                    StorageState:checkStorage(room, "State"),
-                    StorageCapacityUsed:checkStorage(room, "CapacityUsed"),
-                    StorageCapacityFree:checkStorage(room, "CapacityFree"),
-                    StorageCapacityEnergy:checkStorage(room, "CapacityEnergy"),
-                },
+                Storage:CheckRoom.Storage(room),
+                HostileCreeps:CheckRoom.HostileCreeps(room),
+                Sources:CheckRoom.Sources(room),
+                Spawns:CheckRoom.Spawns(room),
+                Minerals:CheckRoom.Minerals(room),
+                Other:CheckRoom.OtherInfo(room),
+                Controller:CheckRoom.Controller(room),
+                Terminal:CheckRoom.Terminal(room)
             },
         }